fix api add request rest

+++ src/main/java/com/example/itspower/response/user/GroupResponse.java
@@ -10,7 +10,7 @@ public class GroupResponse {
     private String groupName = "";
     private Integer groupId;
     private Integer parentId;
-    private Integer demarcationAvailable;
+    private Float demarcationAvailable;
 
     public GroupResponse(GroupEntity groupEntity) {
         this.groupName = groupEntity.getGroupName();
